feat: abstract model and remove context  (#16)
Quite a few changes, but here's breakdown:

- Removed context and made it flat with Workflow
- Removed workflow.message.push in favor of passing messages as second
argument in recursion. This will make workflow immutable and less stable
in case it goes in parallel
- Moved selectAgent and getNextTask to supervisor 
- Moved remaining files from index to teamwork (this is where main
execution happens)
