+++ packages/framework/src/executor.ts
@@ -1,42 +1,9 @@
 import s from 'dedent'
-import OpenAI from 'openai'
 import { zodFunction, zodResponseFormat } from 'openai/helpers/zod'
 import { z } from 'zod'
 
 import { Agent } from './agent.js'
 import { Message } from './types.js'
-
-const openai = new OpenAI()
-
-import { Workflow } from './workflow.js'
-
-// tbd: add more context like trace, callstack etc. context should be serializable
-type ContextOptions = {
-  workflow: Workflow
-  // tbd: move messages to something such as memory
-  messages?: Message[]
-}
-
-/**
- * Helper utility to create a context with defaults.
- */
-export const context = (options: ContextOptions): Context => {
-  return {
-    ...options,
-    messages: [
-      {
-        role: 'assistant',
-        content: s`
-        Here is description of the workflow and expected output by the user:
-        <workflow>${options.workflow.description}</workflow>
-        <output>${options.workflow.output}</output>
-      `,
-      },
-    ],
-  }
-}
-
-export type Context = Required<ContextOptions>
 
 // tbd: helper utilities to create contexts from workflows with concrete single task etc.
 
@@ -57,22 +24,21 @@ export async function executeTaskWithAgent(
         )
       : []
 
-    const response = await openai.beta.chat.completions.parse({
-      model: agent.model as string,
+    const response = await agent.provider.completions({
       // tbd: verify the prompt
       messages: [
         {
           role: 'system',
           content: s`
-                You are ${agent.role}. ${agent.description}
-                
-                Your job is to complete the assigned task.
-                1. You can break down the task into steps
-                2. You can use available tools when needed
-    
-                First try to complete the task on your own.
-                Only ask question to the user if you cannot complete the task without their input.
-              `,
+            You are ${agent.role}. ${agent.description}
+            
+            Your job is to complete the assigned task.
+            1. You can break down the task into steps
+            2. You can use available tools when needed
+
+            First try to complete the task on your own.
+            Only ask question to the user if you cannot complete the task without their input.
+          `,
         },
         ...messages,
       ],
