+++ packages/framework/src/types.ts
@@ -1,3 +1,5 @@
+import { ChatCompletionMessageParam } from 'openai/resources/index.mjs'
+
 /**
  * Utility type to get optional keys from T.
  */
@@ -14,3 +16,5 @@ export type OptionalProperties<T> = Pick<T, OptionalKeys<T>>
  * Utility type to make optional properties required (only includes optional props).
  */
 export type RequiredOptionals<T> = Required<OptionalProperties<T>>
+
+export type Message = ChatCompletionMessageParam
